Update road-to-websockets.md
My edits probably need to be improved but I really think that the earlier history of sockets is important. Web Sockets is a protocol that builds on sockets (right?) and the networking protocols that the WorldWideWeb program used that have existed for a much longer time. I think there are other inconsistencies in this article that need reconciliation.